datahubterminalplugin: Fix warning in Terminal::publish method. WIP fixing mounting of structs as components and definition of inner structs. Fix DatamartDefinition entity and struct default methods.

+++ test/test/org/example/test/Server.java
@@ -25,7 +25,9 @@ public class Server {
 	}
 
 	private static void normalizeTimelineExtensions() {
-		var files = FileUtils.listFiles(new File("C:\\Users\\naits\\Desktop\\IntinoDev\\ness\\temp\\datahub\\datamarts\\master\\timelines"), new String[]{"tl"}, true);
+		File directory = new File("C:\\Users\\naits\\Desktop\\IntinoDev\\ness\\temp\\datahub\\datamarts\\master\\timelines");
+		if(!directory.exists()) return;
+		var files = FileUtils.listFiles(directory, new String[]{"tl"}, true);
 		for(File file : files) {
 			file.renameTo(new File(file.getAbsolutePath().replace(".tl", ".timeline")));
 		}
